User fixes and additions

+++ modules/user/classes/controller/admin/user.php
@@ -4,7 +4,7 @@
  *
  * @package   Gleez\User\Admin\Controller
  * @author    Gleez Team
- * @version   1.0.2
+ * @version   1.0.4
  * @copyright (c) 2011-2013 Gleez Technologies
  * @license   http://gleezcms.org/license
  */
@@ -81,6 +81,12 @@ class Controller_Admin_User extends Controller_Admin {
 
 	/**
 	 * Add new user
+	 *
+	 * @uses  Message::success
+	 * @uses  Route::get
+	 * @uses  Route::uri
+	 * @uses  Arr::merge
+	 * @uses  Arr::get
 	 */
 	public function action_add()
 	{
@@ -100,39 +106,51 @@ class Controller_Admin_User extends Controller_Admin {
 
 		if ($this->valid_post('user'))
 		{
-			try
+			$data = Validation::factory($this->request->post())
+				->rule('pass', 'not_empty')
+				->rule('pass', 'min_length', array(':value', Config::get('auth.password.length_min', 4)))
+				->label('pass', __('Password'));
+
+			if ($data->check())
 			{
-				// Affects the sanitized vars to the user object
-				$post->values($_POST);
-
-				// Create the User
-				$post->save();
-
-				// Make sure to add an empty if none of the roles checked to avoid errors
-				if (empty($_POST['roles']))
+				try
 				{
-					$_POST['roles'] = array('login');
-				}
+					// Affects the sanitized vars to the user object
+					$post->values($this->request->post());
 
-				// Make sure to add an empty if none of the roles checked to avoid errors
-				if (empty($_POST['roles']) OR ! in_array('login', $_POST['roles']))
+					// Create the User
+					$post->save();
+
+					// Make sure to $_POST is set and it is array
+					if ( ! isset($_POST['roles']) OR ! is_array($_POST['roles']))
+					{
+						$_POST['roles'] = array();
+					}
+
+					// Make sure to add an empty if none of the roles checked to avoid errors
+					if (empty($_POST['roles']) OR is_null(Arr::get($_POST['roles'], 'login', NULL)))
+					{
+						$_POST['roles'] = Arr::merge($_POST['roles'], array('login' => ''));
+					}
+
+					foreach(array_keys($_POST['roles']) as $role)
+					{
+						// add() executes the query immediately, and saves the data
+						$post->add('roles', ORM::factory('role', array('name' => $role)));
+					}
+
+					Message::success(__("User %name saved successful!", array('%name' => $post->name)));
+
+					$this->request->redirect(Route::get('admin/user')->uri(), 200);
+				}
+				catch (ORM_Validation_Exception $e)
 				{
-					$_POST['roles'] = array();
+					$this->_errors = $e->errors('models', TRUE);
 				}
-
-				foreach(array_keys($_POST['roles']) as $role)
-				{
-					// add() executes the query immediately, and saves the data
-					$post->add('roles', ORM::factory('role', array('name' => $role)));
-				}
-
-				Message::success(__("User %name saved successful!", array('%name' => $post->name)));
-
-				$this->request->redirect(Route::get('admin/user')->uri(array('action' => 'list')), 200);
 			}
-			catch (ORM_Validation_Exception $e)
+			else
 			{
-				$this->_errors = $e->errors('models', TRUE);
+				$this->_errors = $data->errors('models', TRUE);
 			}
 		}
 
@@ -141,6 +159,14 @@ class Controller_Admin_User extends Controller_Admin {
 
 	/**
 	 * Edit user
+	 *
+	 * @uses  Message::error
+	 * @uses  Message::success
+	 * @uses  Log::error
+	 * @uses  Route::get
+	 * @uses  Route::uri
+	 * @uses  Arr::merge
+	 * @uses  Arr::get
 	 */
 	public function action_edit()
 	{
@@ -173,41 +199,59 @@ class Controller_Admin_User extends Controller_Admin {
 
 		if ($this->valid_post('user'))
 		{
-			try
+			$data = Validation::factory($this->request->post())
+				->rule('pass', 'not_empty')
+				->rule('pass', 'min_length', array(':value', Config::get('auth.password.length_min', 4)))
+				->label('pass', __('Password'));
+
+			if ($data->check())
 			{
-				// password can be empty - it will be ignored in save.
-				if ((empty($_POST['pass']) || (trim($_POST['pass']) == '')))
+				try
 				{
-					unset($_POST['pass']);
+					// password can be empty - it will be ignored in save.
+					if ((empty($_POST['pass']) || (trim($_POST['pass']) == '')))
+					{
+						unset($_POST['pass']);
+					}
+
+					$post->values($this->request->post());
+					$post->save();
+
+					// Make sure to $_POST is set and it is array
+					if ( ! isset($_POST['roles']) OR ! is_array($_POST['roles']))
+					{
+						$_POST['roles'] = array();
+					}
+
+					// Make sure to add an empty if none of the roles checked to avoid errors
+					if (empty($_POST['roles']) OR is_null(Arr::get($_POST['roles'], 'login', NULL)))
+					{
+						$_POST['roles'] = Arr::merge($_POST['roles'], array('login' => ''));
+					}
+
+					// Roles have to be added separately, and all users have to have the login role
+					// you first have to remove the items, otherwise add() will try to add duplicates
+					// could also use array_diff, but this is much simpler
+					DB::delete('roles_users')->where('user_id', '=', $id)->execute();
+
+					foreach(array_keys($_POST['roles']) as $role)
+					{
+						// add() executes the query immediately, and saves the data
+						$post->add('roles', ORM::factory('role', array('name' => $role)));
+					}
+
+					Message::success(__("User %name saved successful!", array('%name' => $post->name)));
+
+					$this->request->redirect(Route::get('admin/user')->uri(), 200);
 				}
-
-				$post->values($_POST);
-				$post->save();
-
-				// Make sure to add an empty if none of the roles checked to avoid errors
-				if (empty($_POST['roles']) OR ! in_array('login', $_POST['roles']))
+				catch (ORM_Validation_Exception $e)
 				{
-					$_POST['roles'] = array('login');
+					$this->_errors = $e->errors('models', TRUE);
 				}
-
-				// Roles have to be added separately, and all users have to have the login role
-				// you first have to remove the items, otherwise add() will try to add duplicates
-				// could also use array_diff, but this is much simpler
-				DB::delete('roles_users')->where('user_id', '=', $id)->execute();
-
-				foreach(array_keys($_POST['roles']) as $role)
-				{
-					// add() executes the query immediately, and saves the data
-					$post->add('roles', ORM::factory('role', array('name' => $role)));
-				}
-
-				Message::success(__("User %name saved successful!", array('%name' => $post->name)));
-
-				$this->request->redirect(Route::get('admin/user')->uri());
 			}
-			catch (ORM_Validation_Exception $e)
+			else
 			{
-				$this->_errors = $e->errors('models', TRUE);
+				$this->_errors = $data->errors('models', TRUE);
 			}
 		}
